# api_server/contracts/openapi.v1.json
{
  "components": {
    "responses": {
      "ApiError": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiErrorEnvelope"
            }
          }
        },
        "description": "Typed, content-free API error"
      }
    },
    "schemas": {
...
      "ApiErrorEnvelope": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "minLength": 1,
            "title": "Code",
            "type": "string"
          },
          "message": {
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "format": "uuid",
            "minLength": 1,
            "title": "Request Id",
            "type": "string"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message",
          "retryable",
          "request_id"
        ],
        "title": "ApiErrorEnvelope",
        "type": "object"
      },
...
      "ConfirmRegionsRequest": {
        "additionalProperties": false,
        "properties": {
          "job_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          },
          "job_revision": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Revision"
          },
          "revision": {
            "minimum": 1.0,
            "title": "Revision",
            "type": "integer"
          }
        },
        "required": [
          "revision"
        ],
        "title": "ConfirmRegionsRequest",
        "type": "object"
      },
...
      "Point": {
        "additionalProperties": false,
        "properties": {
          "x": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "X",
            "type": "number"
          },
          "y": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Y",
            "type": "number"
          }
        },
        "required": [
          "x",
          "y"
        ],
        "title": "Point",
        "type": "object"
      },
      "PolygonPoint": {
        "additionalProperties": false,
        "properties": {
          "x": {
            "maximum": 1,
            "minimum": 0,
            "title": "X",
            "type": "number"
          },
          "y": {
            "maximum": 1,
            "minimum": 0,
            "title": "Y",
            "type": "number"
          }
        },
        "required": [
          "x",
          "y"
        ],
        "title": "PolygonPoint",
        "type": "object"
      },
...
      "RecognitionRegion": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "page_id": {
            "format": "uuid",
            "title": "Page Id",
            "type": "string"
          },
          "polygon": {
            "items": {
              "$ref": "#/components/schemas/PolygonPoint"
            },
            "minItems": 4,
            "title": "Polygon",
            "type": "array"
          },
          "position": {
            "minimum": 0,
            "title": "Position",
            "type": "integer"
          },
          "revision": {
            "minimum": 1,
            "title": "Revision",
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "created_at",
          "updated_at",
          "page_id",
          "polygon",
          "position",
          "revision"
        ],
        "title": "RecognitionRegion",
        "type": "object"
      },
      "RegionCreate": {
        "additionalProperties": false,
        "properties": {
          "idempotency_key": {
            "maxLength": 128,
            "minLength": 16,
            "title": "Idempotency Key",
            "type": "string"
          },
          "page_id": {
            "format": "uuid",
            "title": "Page Id",
            "type": "string"
          },
          "polygon": {
            "items": {
              "$ref": "#/components/schemas/PolygonPoint"
            },
            "minItems": 4,
            "title": "Polygon",
            "type": "array"
          },
          "position": {
            "minimum": 0,
            "title": "Position",
            "type": "integer"
          }
        },
        "required": [
          "page_id",
          "polygon",
          "position",
          "idempotency_key"
        ],
        "title": "RegionCreate",
        "type": "object"
      },
      "RegionInput": {
        "additionalProperties": false,
        "properties": {
          "detector_score": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detector Score"
          },
          "detector_version": {
            "anyOf": [
              {
                "maxLength": 128,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detector Version"
          },
          "flags": {
            "items": {
              "type": "string"
            },
            "maxItems": 32,
            "title": "Flags",
            "type": "array"
          },
          "id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "polygon": {
            "items": {
              "$ref": "#/components/schemas/Point"
            },
            "maxItems": 4,
            "minItems": 4,
            "title": "Polygon",
            "type": "array"
          },
          "reading_order": {
            "maximum": 100000.0,
            "minimum": 0.0,
            "title": "Reading Order",
            "type": "integer"
          },
          "source": {
            "enum": [
              "craft",
              "manual",
              "adjusted"
            ],
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "polygon",
          "reading_order",
          "source"
        ],
        "title": "RegionInput",
        "type": "object"
      },
...
      "RegionResponse": {
        "properties": {
          "detector_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detector Score"
          },
          "detector_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detector Version"
          },
          "flags": {
            "items": {
              "type": "string"
            },
            "title": "Flags",
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "polygon": {
            "items": {
              "$ref": "#/components/schemas/Point"
            },
            "title": "Polygon",
            "type": "array"
          },
          "reading_order": {
            "title": "Reading Order",
            "type": "integer"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "source": {
            "enum": [
              "craft",
              "manual",
              "adjusted"
            ],
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "id",
          "polygon",
          "reading_order",
          "revision",
          "source",
          "flags",
          "detector_version",
          "detector_score"
        ],
        "title": "RegionResponse",
        "type": "object"
      },
      "RegionsResponse": {
        "properties": {
          "confirmed": {
            "title": "Confirmed",
            "type": "boolean"
          },
          "page_id": {
            "format": "uuid",
            "title": "Page Id",
            "type": "string"
          },
          "regions": {
            "items": {
              "$ref": "#/components/schemas/RegionResponse"
            },
            "title": "Regions",
            "type": "array"
          },
          "resumed_job_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resumed Job Id"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          }
        },
        "required": [
          "page_id",
          "revision",
          "confirmed",
          "regions"
        ],
        "title": "RegionsResponse",
        "type": "object"
      },
      "ReplaceRegionsRequest": {
        "additionalProperties": false,
        "properties": {
          "regions": {
            "items": {
              "$ref": "#/components/schemas/RegionInput"
            },
            "maxItems": 10000,
            "title": "Regions",
            "type": "array"
          },
          "revision": {
            "minimum": 1.0,
            "title": "Revision",
            "type": "integer"
          }
        },
        "required": [
          "revision",
          "regions"
        ],
        "title": "ReplaceRegionsRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "CookieSession": {
        "description": "HttpOnly same-origin session cookie",
        "in": "cookie",
        "name": "htr_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Versioned API v1 contract. Operations marked planned are contracts only and are not exposed by the runtime until implemented.",
    "title": "Tajik HTR Studio API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/pages/{page_id}/regions": {
      "get": {
        "operationId": "liveGetApiV1PagesPageIdRegionsV1",
        "parameters": [
          {
            "in": "path",
            "name": "page_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegionsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "$ref": "#/components/responses/ApiError"
          },
          "default": {
            "$ref": "#/components/responses/ApiError"
          }
        },
        "security": [
          {
            "CookieSession": []
          }
        ],
        "summary": "Get Regions",
        "tags": [
          "regions"
        ],
        "x-implementation-status": "live"
      },
      "post": {
        "operationId": "createPageRegionV1",
        "parameters": [
          {
            "in": "path",
            "name": "page_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecognitionRegion"
                }
              }
            },
            "description": "Successful response"
          },
          "default": {
            "$ref": "#/components/responses/ApiError"
          }
        },
        "security": [
          {
            "CookieSession": []
          }
        ],
        "summary": "Create a manual recognition region",
        "tags": [
          "regions"
        ],
        "x-implementation-status": "planned"
      },
      "put": {
        "operationId": "livePutApiV1PagesPageIdRegionsV1",
        "parameters": [
          {
            "in": "path",
            "name": "page_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplaceRegionsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegionsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "$ref": "#/components/responses/ApiError"
          },
          "default": {
            "$ref": "#/components/responses/ApiError"
          }
        },
        "security": [
          {
            "CookieSession": []
          }
        ],
        "summary": "Replace Regions",
        "tags": [
          "regions"
        ],
        "x-implementation-status": "live"
      }
    },
    "/api/v1/pages/{page_id}/regions/confirm": {
      "post": {
        "operationId": "livePostApiV1PagesPageIdRegionsConfirmV1",
        "parameters": [
          {
            "in": "path",
            "name": "page_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmRegionsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegionsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "$ref": "#/components/responses/ApiError"
          },
          "default": {
            "$ref": "#/components/responses/ApiError"
          }
        },
        "security": [
          {
            "CookieSession": []
          }
        ],
        "summary": "Confirm Regions",
        "tags": [
          "regions"
        ],
        "x-implementation-status": "live"
      }
    },
...
  },
  "servers": [
    {
      "description": "Same-origin deployment",
      "url": "/"
    }
  ],
...
}